unequal list order check
